avatar/requirements.txt
azure-cognitiveservices-speech
azure-identity
flask
openai
pytz
requests
sseclient-py
Flask-Session
Flask-Session[cachelib]